+++ src/main/java/insane96mcp/insanesurvivaloverhaul/mixin/accessor/MobEffectInstanceAccessor.java
@@ -1,0 +1,13 @@
+package insane96mcp.insanesurvivaloverhaul.mixin.accessor;
+
+import net.minecraft.world.effect.MobEffectInstance;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Mutable;
+import org.spongepowered.asm.mixin.gen.Accessor;
+
+@Mixin(MobEffectInstance.class)
+public interface MobEffectInstanceAccessor {
+    @Mutable
+    @Accessor
+    void setDuration(int duration);
+}